added postMultiply functions (#657)
For every existing multiply*() function, such as multiplyCompMatr(), this commit adds a corresponding postMultiply*() function which operates upon a density matrix from the right-hand side. This is useful for preparing density matrices in non-physical states which appear as sub-expressions within things like commutators and the Linbladian. 

Implementing these functions involved:
- updating the templating of "any target dense matrix" function, inadvertently simplifying the associated instantiation and dispatch macros by re-using those for the "any target diagonal matrix" function
- adding new utilities and logic for obtaining/effecting the transpose of a function, to undo the transpose effected via operation upon the bra-qubits of a vectorised density matrix
- extending and refactoring the unit tests with postMultiply references

We additionally added the below expected but missing functions from the API:
- multiplyPauliX
- multiplyPauliY
- multiplyPauliZ